requirements.txt
alembic==1.4.2
autopep8==1.5.3
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==2.0.0a1
Jinja2==3.0.0a1
Mako==1.1.3
MarkupSafe==2.0.0a1
mysql-connector-python==8.0.21
Pillow==7.2.0
protobuf==4.0.0rc1
pycodestyle==2.6.0
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
Werkzeug==1.0.1
WTForms==2.3.1
